requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
bcrypt==4.1.3
binaryornot==0.4.4
capstone==5.0.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmd2==2.4.3
colored-traceback==0.4.2
colorlog==6.8.2
cookiecutter==2.6.0
cryptography==43.0.0
Django==4.2.14
django-cors-headers==4.4.0
django-extensions==3.2.3
exceptiongroup==1.2.2
fastapi==0.112.1
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
intervaltree==3.1.0
jinja2==3.1.4
linkify-it-py==2.0.3
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
netifaces==0.11.0
oauthlib==3.2.2
packaging==23.2
paramiko==3.4.0
pluggy==1.5.0
plumbum==1.8.3
psutil==6.0.0
pwnlib==8.6.8.6
pwntools==4.12.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pyelftools==0.31
pygments==2.18.0
PyNaCl==1.5.0
pyperclip==1.9.0
pypng==0.20220715.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pyusb==1.2.1
pywifi==1.1.12
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==13.7.1
ROPGadget==7.4
rpyc==6.0.0
sentry-sdk==2.10.0
six==1.16.0
smbus==1.1.post2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
starlette==0.38.2
text-unidecode==1.3
textual==0.77.0
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.2
uc-micro-py==1.0.3
unicorn==2.0.1.post1
unix-ar==0.2.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.23.0
wcwidth==0.2.13
websockets==12.0
zstandard==0.23.0
